+++ utils/templates/validation/builder/Elk_Aerial_General_Recruit_Comp_Survey_2.0_Validation_Config_Builder.ts
@@ -15,7 +15,7 @@ import {
   surveyOrTelemetryPickListValidator,
   targetPickListValidator,
   utmZoneValidator
-} from '../../validation-config-helpers';
+} from '../../helper/validation-config-helpers';
 
 const elkSpeciesValidator = () => {
   return [
