+++ utils/templates/validation/builder/Elk_Aerial_SRB_Recruit_Comp_Survey_2.0_Validation_Config_Builder.ts
@@ -13,7 +13,7 @@ import {
   surveyOrTelemetryPickListValidator,
   targetPickListValidator,
   utmZoneValidator
-} from '../../validation-config-helpers';
+} from '../../helper/validation-config-helpers';
 
 const elkSpeciesValidator = () => {
   return [
